+++ cpm/generators/simulator.py
@@ -27,8 +27,8 @@ class Simulator:
 
     Attributes
     ----------
-    function : object
-        The simulation function to be used.
+    wrapper : Wrapper
+        The simulation wrapper to be used.
     data : object
         The data required for the simulation.
     parameters : object
@@ -47,8 +47,8 @@ class Simulator:
 
     """
 
-    def __init__(self, model=None, data=None, parameters=None):
-        self.function = model
+    def __init__(self, wrapper=None, data=None, parameters=None):
+        self.wrapper = wrapper
         self.data = data
         self.parameters = copy.deepcopy(parameters)
         if len(self.data) != len(self.parameters):
@@ -59,7 +59,7 @@ class Simulator:
             warnings.warn(
                 "The number of parameter sets and number of participants in data do not match.\nUsing the same parameters for all participants."
             )
-        self.parameter_names = self.function.parameter_names
+        self.parameter_names = self.wrapper.parameter_names
         self.simulation = []
         self.generated = []
         self.__run__ = False
@@ -73,8 +73,8 @@ class Simulator:
         experiment: A list containing the results of the simulation.
         """
         for i in range(len(self.data)):
-            self.function.reset()
-            evaluate = copy.deepcopy(self.function)
+            self.wrapper.reset()
+            evaluate = copy.deepcopy(self.wrapper)
             evaluate.data = self.data[i]
             evaluate.reset(parameters=self.parameters[i])
             evaluate.run()
@@ -118,21 +118,27 @@ class Simulator:
             self.parameters = parameters
         return None
 
-    def generate(self):
+    def generate(self, variable="dependent"):
         """
         Generate data for parameter recovery, etc.
+
+        Parameters
+        ----------
+        variable: str
+            Name of the variable to pull out from model output.
 
         Returns
         ------
         results: numpy.ndarray
             An array of dictionaries containing the results of the simulation.
         """
-        for i in self.simulation:
-            current = []
-            for j in i:
-                current.append(j.get("dependent"))
-            self.generated.append({"observed": np.asarray(current)})
-        self.generated = np.array(self.generated)
+        append = []
+        for ppt in self.simulation:
+            one = {"observed": np.zeros((self.wrapper.__len__, 1))}
+            for k in range(self.wrapper.__len__):
+                one.get("observed")[k] = np.array([ppt[k].get(variable)])
+            append.append(one)
+        self.generated = copy.deepcopy(append)
         return None
 
     def reset(self):